chore: restrict the cond of decoding key to basic types for mutations
Only when a basic type / any type is annotated, we decode the primary
key columns to a single value instead of a struct.

This means they can always explicitly create a struct type (namedtuple
or dataclass) if they want even when there's only a single primary key
column. This will allow users to handle multiple-primary-key-columns in
a consistent way.